commit rest of code

+++ server/src/routes/removeProgressRoute.js
@@ -2,7 +2,9 @@ const ProgressModel = require('../models/progressModel');
 module.exports = async (req, res) => {
     const { id } = req.params;
     const progress = await ProgressModel.findById(id);
-    await progress.deleteOne();
+    if(progress!=null){
+        await progress.deleteOne();
+    }
     res.status(204).json(progress);
 
 }
